requirements.txt
streamlit==1.30.0
pandas==2.2.1
plotly==5.20.0
passlib==1.7.4
supabase==1.0.3
python-dotenv==1.0.1
psycopg2-binary==2.9.9

